requirements.txt
glfw>=2.5.9
imgui[glfw]>=2.0.0
numpy>=1.24.2
PyOpenGL>=3.1.6
